requirements.txt
streamlit
openai
pandas
pydanticsendgrid
requests
beautifulsoup4
